Adicionando o tratamento de excessoes com RestControllerAdvice

+++ src/main/java/com/github/delduck/spring_product_api/service/ProdutoService.java
@@ -1,5 +1,6 @@
 package com.github.delduck.spring_product_api.service;
 
+import com.github.delduck.spring_product_api.exceptions.RecursoNaoEncontratoException;
 import com.github.delduck.spring_product_api.model.Produto;
 import com.github.delduck.spring_product_api.repository.ProdutoRepository;
 import org.springframework.stereotype.Service;
@@ -20,8 +21,9 @@ public class ProdutoService {
         return produtoRepository.findAll();
     }
 
-    public Optional<Produto> buscarPorId(Long id) {
-        return produtoRepository.findById(id);
+    public Produto buscarPorId(Long id) {
+        return produtoRepository.findById(id)
+                .orElseThrow(() -> new RecursoNaoEncontratoException("Produto com ID " + id + " não encontrado."));
     }
 
     public Produto salvarProduto(Produto produto) {
@@ -29,6 +31,9 @@ public class ProdutoService {
     }
 
     public void deletarProduto(Long id) {
+        if(!produtoRepository.existsById(id)) {
+            throw new RecursoNaoEncontratoException("Produto com ID " + id + " não encontrado.");
+        }
         produtoRepository.deleteById(id);
     }
